Eliminate stdout from IO#dup spec runs.
At least, under JRuby it was printing things like:
"No such file or directory".

+++ spec/ruby/1.8/core/io/dup_spec.rb
@@ -21,7 +21,7 @@ describe "IO#dup" do
   end
 
   after :all do
-    File.unlink @file
+    File.unlink @file if File.exists?(@file)
   end
 
   it "returns a new IO instance" do
